backend/requirements.txt
fastapi
uvicorn[standard]
pymongo[srv]==3.12
python-dotenv
pydantic
pillow